Improve web models documentation
